backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
requests==2.32.5
mongoengine==0.29.1
pymongo==4.16.0
dnspython==2.8.0
djangorestframework-simplejwt==5.3.1

